LibEDID: Export common structures into a header file
This will be used later in the kernel code to simplify the EDID handling
code.
